Refactored PgnTable
