+++ src/features/panel/Panel.js
@@ -13,8 +13,7 @@ import StartedButtonsStockfishMode from 'features/mode/stockfish/StartedButtonsS
 import Timer from 'features/mode/play/timer/Timer';
 import History from 'features/panel/History';
 import OpeningTable from 'features/panel/OpeningTable';
-import RavTable from 'features/panel/RavTable';
-import SanTable from 'features/panel/SanTable';
+import PgnTable from 'features/panel/PgnTable';
 import PgnButtons from 'features/panel/PgnButtons';
 import PgnDialogs from 'features/panel/PgnDialogs';
 
@@ -45,7 +44,7 @@ const Panel = ({ props }) => {
             <History />
           </Grid>
           <Grid item xs={12}>
-            <RavTable />
+            <PgnTable />
           </Grid>
         </Grid>
         <Grid item xs={12} sx={styles.buttons}>
